pyproject.toml
[tool.poetry]
name = "llm-student-onboarding"
version = "0.1.0"
description = ""
authors = ["Satya Adda"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
langchain = "^0.1.11"
pandas = "^2.2.1"
langchain-openai = "^0.0.8"
langchain-community = "^0.0.27"
python-dotenv = "^1.0.1"
chromadb = "^0.4.24"
pypdf = "^4.1.0"
llama-index = "^0.10.20"
unstructured = {extras = ["docx"], version = "^0.12.6"}
duckduckgo-search = "^5.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
